requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
Flask-PyMongo==3.0.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
pika==1.3.2
pymongo==4.11.2
python-decouple==3.8
requests==2.32.3
urllib3==2.3.0
Werkzeug==3.1.3
